implement admin system

+++ src/main/java/com/elec5619/bloodsystem/service/AccountService.java
@@ -22,7 +22,7 @@ public class AccountService {
 
 
 
-    public Account getAccountById(int id){
+    public Account getAccountById(Long id){
         return accountRepository.findById(id);
     }
